+++ docs/src/TextArea.doc.js
@@ -1,8 +1,7 @@
 // @flow
 import * as React from 'react';
-import { Box, Label, TextArea } from 'gestalt';
+import Example from './components/Example';
 import PropTable from './components/PropTable';
-import StateRecorder from './components/StateRecorder';
 import PageHeader from './components/PageHeader';
 import Card from './components/Card';
 import CardPage from './components/CardPage';
@@ -77,106 +76,129 @@ card(
 );
 
 card(
-  <StateRecorder
+  <Example
+    name="Example"
     description={`
     A \`TextArea\` will expand to fill the width of their parent container.
-
-    ~~~jsx
-    <Box marginBottom={2}>
-      <Label htmlFor="aboutme">With a placeholder</Label>
-    </Box>
-    <TextArea id="aboutme" placeholder="Write something about yourself..." />
-    ~~~
   `}
-    name="Options"
-    fn={atom => (
-      <Box paddingX={2}>
-        <Box paddingY={3}>
-          <Box marginBottom={2}>
-            <Label htmlFor="aboutme">With a placeholder</Label>
-          </Box>
-          <TextArea
-            id="aboutme"
-            onChange={({ value }) => atom.reset({ placeholderValue: value })}
-            placeholder="Write something about yourself..."
-            value={atom.deref().placeholderValue}
-          />
-        </Box>
+    defaultCode={`
+class Example extends React.Component {
+  constructor(props) {
+    super(props);
+    this.handleChange = this._handleChange.bind(this);
+    this.state = {
+      value: ""
+    };
+  }
+  _handleChange({ value }) {
+    this.setState({
+      value
+    });
+  }
 
-        <Box paddingY={3}>
-          <Box marginBottom={2}>
-            <Label htmlFor="disabled">Disabled</Label>
-          </Box>
-          <TextArea
-            disabled
-            id="disabled"
-            onChange={({ value }) => atom.reset({ disabled: value })}
-            placeholder="A disabled TextArea"
-            value={atom.deref().disabled}
-          />
+  render() {
+    return (
+      <Box>
+        <Box marginBottom={2}>
+          <Label htmlFor="aboutme">
+            <Text>With a placeholder</Text>
+          </Label>
         </Box>
+        <TextArea
+          id="aboutme"
+          onChange={this.handleChange}
+          placeholder="Write something about yourself..."
+          value={this.state.value}
+        />
       </Box>
-    )}
-    historyLimit={100}
+    );
+  }
+}
+`}
   />
 );
 
 card(
-  <StateRecorder
-    description={`
-    TextArea's can display their own error messages if you'd like them to.
-    To use our errors, simply pass in an \`errorMessage\` when there is an error present and we will
-    handle the rest.
+  <Example
+    name="Example: Disabled"
+    defaultCode={`
+class Example extends React.Component {
+  constructor(props) {
+    super(props);
+    this.handleChange = this._handleChange.bind(this);
+    this.state = {
+      value: ""
+    };
+  }
+  _handleChange({ value }) {
+    this.setState({
+      value
+    });
+  }
 
-    ~~~jsx
-    <TextArea errorMessage="This field can't be blank!" id="comment" />
-    ~~~
-  `}
-    name="Errors"
-    fn={atom => (
-      <Box paddingX={2}>
-        <Box paddingY={3}>
-          <Box marginBottom={2}>
-            <Label htmlFor="comment">With error message</Label>
-          </Box>
-          <TextArea
-            errorMessage="This field can't be blank!"
-            id="comment"
-            onChange={({ value }) => atom.reset({ errorMsgValue: value })}
-            value={atom.deref().errorMsgValue}
-          />
+  render() {
+    return (
+      <Box>
+        <Box marginBottom={2}>
+          <Label htmlFor="aboutme">
+            <Text>Disabled</Text>
+          </Label>
         </Box>
+        <TextArea
+          disabled
+          id="aboutme"
+          onChange={this.handleChange}
+          placeholder="Write something about yourself..."
+          value={this.state.value}
+        />
       </Box>
-    )}
+    );
+  }
+}
+`}
   />
 );
 
 card(
-  <StateRecorder
+  <Example
+    name="Example: Error message"
     description={`
-    TextArea's can resize their height to fit different line counts besides the default 3.
-    To use non-default height, pass in \`rows\` to set the underlying value on the html element.
+    A TextArea can display its own error message.
+    To use our errors, simply pass in an \`errorMessage\` when there is an error present and we will     handle the rest.`}
+    defaultCode={`
+class Example extends React.Component {
+  constructor(props) {
+    super(props);
+    this.handleChange = this._handleChange.bind(this);
+    this.state = {
+      value: ""
+    };
+  }
+  _handleChange({ value }) {
+    this.setState({
+      value
+    });
+  }
 
-    ~~~jsx
-    <TextArea id="rows" rows={5} />
-    ~~~
-  `}
-    name="Rows"
-    fn={atom => (
-      <Box paddingX={2}>
-        <Box paddingY={3}>
-          <Box marginBottom={2}>
-            <Label htmlFor="rows">With non-default rows</Label>
-          </Box>
-          <TextArea
-            id="rows"
-            onChange={({ value }) => atom.reset({ placeholderValue: value })}
-            rows={5}
-            value={atom.deref().placeholderValue}
-          />
+  render() {
+    return (
+      <Box>
+        <Box marginBottom={2}>
+          <Label htmlFor="aboutme">
+            <Text>With an error message</Text>
+          </Label>
         </Box>
+        <TextArea
+          id="aboutme"
+          errorMessage={!this.state.value ? "This field can't be blank!" : null}
+          onChange={this.handleChange}
+          value={this.state.value}
+        />
       </Box>
-    )}
+    );
+  }
+}
+`}
   />
 );